requirements.txt
asgiref==3.8.1
coverage==7.6.4
Django==5.1.1
django-authtools==2.0.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
ruff==0.7.1
sqlparse==0.5.1
tzdata==2024.1
